requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.4
PyJWT==1.7.1
pylint==2.4.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2